% File man/valorate-internal.Rd

\name{valorate-internal}

%% Not documented
%% <LIST>
\alias{valorate.p.overlap}
\alias{valorate.hyper.density}
\alias{valorate.comb}
\alias{valorate.perm}
\alias{valorate.cat}
\alias{valorate.mav}
\alias{plot_kaplan_valorate}
\alias{valorate.estimate.beta.parameters}
\alias{valorate.estimate.weibull.parameters}
\alias{valorate.psurvdiff}
\alias{valorate.plot.diff.empirical}
%% </LIST>

\title{Internal VALORATE Functions}
\description{
  Internal Valorate functions
}
\details{
  Not to be called by user.
}
\keyword{ internal }
